remove extra field, add @Builder.Default, move CalendarService.java to config and rename one to CalendarConfig
