--- src/main/java/by/exadel/internship/service/CalendarService.java
@@ -1,7 +1,0 @@
-package by.exadel.internship.service;
-
-import com.google.api.services.calendar.Calendar;
-
-public interface CalendarService {
-    Calendar getDefaultCalendar();
-}
